Add items to Faq
Refactor and separate data from Faq component.
