--- src/Cities.ts
@@ -1,49 +1,0 @@
-export const cities = [
-  {
-    name: 'Stockholm',
-    date: '18.4.2025',
-    imageUrlSmall: '/assets/stockholm_small.webp',
-    imageUrl: '/assets/stockholm.webp',
-    agenda: `Friday, April 18th, 2025
-
-    Run in Stockholm hosted by MRC Stockholm. Time to be confirmed.
-
-    Post-run beers and party.
-
-    Boat 17:30 Tallink Silja Baltic Queen to Tallinn; overnight trip, arrival to Tallinn Saturday 10:45.`,
-  },
-  {
-    name: 'Tallinn',
-    date: '19.4.2025',
-    imageUrlSmall: '/assets/tallinn_small.webp',
-    imageUrl: '/assets/tallinn.webp',
-    agenda: `Saturday, April 19th, 2025
-
-    Run in Tallinn hosted by Pühaste jooksuklubi.
-
-    Post-run beers and party.
-
-    Boat 22:30 Tallink Silja Megastar to Helsinki (2 hour trip, arrives 00:30)
-
-   Note! Feel free to book earlier or later boat to suit your needs eg. depending on choice of hotel and if you want to overnight in Helsinki (recommended for those who want to sleep late) or Tallinn (if you prefer to catch an early boat at around 6 or 7).`,
-  },
-  {
-    name: 'Helsinki',
-    date: '20.4.2025',
-    imageUrlSmall: '/assets/helsinki_small.webp',
-    imageUrl: '/assets/helsinki.webp',
-    agenda: `Sunday, April 20th, 2025
-
-    Run in Helsinki hosted by MRC Helsinki. Starting at 12:30 from SalamaNation.
-
-    Post-run beers and party at SalamaNation.
-
-    Optional beer mile (to be confirmed).
-
-    Awards and farewells until we meet again.
-
-    Suggested boat for Stockholm-bound runners: 17:00 Silja Serenade to Stockholm. Arrives to STO on Monday 10:00.
-    
-    Note! Feel free to book another boat or flight, or stay in town until Monday.`,
-  },
-];
